Fix ClangSharp 21 pipeline integration
Restore SDK-derived DXVA headers and select C++20 only for affected partitions. Install both required SDK generations in CI and explicitly package only runtime tool dependencies.


Copilot-Session: 19e5498a-dfe7-4601-8955-98f33a92582b

+++ generation/WinSDK/RecompiledIdlHeaders/um/dxva.h
@@ -1958,9 +1958,9 @@ typedef enum _DXVA_SampleFormat {
 #define DXVA_ExtractExtColorData(_sf, _Mask, _Shift) \
     (((_sf) & (_Mask)) >> (_Shift))
 
-#define DXVABitMask(__n) (~((~0u) << __n))
+#define DXVABitMask(__n) (~((~0) << __n))
 #define DXVA_ExtColorData_ShiftBase 8
-#define DXVAColorMask(__bits,__base) ((int)(DXVABitMask(__bits) << (__base)))
+#define DXVAColorMask(__bits,__base) (DXVABitMask(__bits) << (__base))
 
 typedef enum _DXVA_VideoTransferFunction
 {
